周期入力のピッカーを分離、デザイン適用
